AG-12693 CR: Join scene/util/bezier.ts & util/bezier.ts
